productList와 productcontroller mapping
