+++ DressHub/src/main/java/com/newoneplus/dresshub/Controller/BasketController.java
@@ -20,6 +20,15 @@ public class BasketController {
         return productService.getBasketList(page-1);
     }
 
+    @PostMapping
+    public Basket create(@RequestBody Basket basket){
+        return productService.createBasket(basket);
+    }
+
+    @PutMapping
+    public void update(@RequestBody Basket basket){
+        productService.createBasket(basket);
+    }
 
 
 }
