+++ DressHub/src/main/java/com/newoneplus/dresshub/Controller/ProductController.java
@@ -17,6 +17,7 @@ import org.springframework.web.multipart.MultipartFile;
 import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
 import java.awt.image.ImageProducer;
+import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -85,11 +86,11 @@ public class ProductController {
         String path = System.getProperty("user.dir") + "/out/main/resources/static/product_image/";
         new File(path).mkdirs(); // 디렉토리 생성
 
-        BufferedImage image = imageProcesser.getMediumImage(productImage.getInputStream());
-        writeImage(filename, path, image, "medium");
+        BufferedImage image2 = imageProcesser.getMediumImage(productImage.getInputStream());
+        writeImage(filename, path, image2, "medium");
 
-        BufferedImage image2 = imageProcesser.getSmallImage(productImage.getInputStream());
-        writeImage(filename, path, image2, "small");
+        BufferedImage image3 = imageProcesser.getSmallImage(productImage.getInputStream());
+        writeImage(filename, path, image3, "small");
 
         productImage.transferTo(new File(path + "origin" + filename));
     }
@@ -99,46 +100,7 @@ public class ProductController {
         ImageIO.write(image2, "jpg", new File(path + sizeFileName));
     }
 
-    @RequestMapping(value = "/thumbUp", method = RequestMethod.POST)
-    public Product likeCreate(@RequestBody Product product, @RequestParam(value = "state") boolean state) throws ClassNotFoundException {
-        //TODO develop이랑 합치면 넣기 !
-        User user = new User();
-        user.setUid("user1");
-        ThumbUp thumbUp = new ThumbUp();
-        thumbUp.setUid(user.getUid());
-        thumbUp.setProduct(product);
-        if (state) {
-            productService.insertThumup(thumbUp);
-        } else {
-            productService.deleteThumup(thumbUp);
-        }
-        return productService.getProduct(product.getId());
 
-    }
-
-
-    @RequestMapping(value = "/thumbUp/search", method = RequestMethod.GET)
-    public Page<Product> getThumbUpProductList(@RequestParam(value = "page", defaultValue = "1") int page) {
-        User user = new User();
-        user.setUid("user1");
-        return productService.getThumbUpProductList(page, user);
-    }
-
-
-    //TODO 나중에 프론트 위임
-    @RequestMapping(value = "/productList", method = RequestMethod.GET)
-    public String productListView(@RequestParam(value = "category", defaultValue = "10") String category, Model model) {
-        model.addAttribute("category", category);
-        return "productList";
-    }
-
-
-    //TODO 나중에 프론트 위임
-    @RequestMapping(value = "/productDetail", method = RequestMethod.GET)
-    public String getProductDetail(@RequestParam(value = "productId") int productId, Model model) throws ClassNotFoundException {
-        model.addAttribute("productId", productId);
-        return "product_details";
-    }
 
 
 }
